Misc package fixes.
etc/conf.d/espeakup.conf needs to be in the backup array.  How did I forget?
This package shouldn't depend on the kernel module package, per the kernel
packaging guidelines.
The upstream source archive has a very permanent download location now.
The homepage should probably be the author's github page for this project.

The old package was built before the svn transition, so it was time for a
rebuild anyway.



git-svn-id: file:///srv/repos/svn-community/svn@13306 9fca08f4-af9d-4005-b8df-a31f2cc04f65