Rebuild for [community] cleanup.
This package would not build with the current version of glibc.  Several
of the programs define an internal function named getline.
These internal definitions have nothing to do with the getline from glibc.
I made a patch.  The name of each internal getline function is prefixed
with the name of the program which uses it.

Also, I changed the dependency on bash to a dependency on sh.  The
shell scripts in the package are free of bashisms, according to the
checkbashisms tool (available in the AUR).

Finally, I made some miscellaneous changes, such as quoting $pkgdir
and $srcdir.



git-svn-id: file:///srv/repos/svn-community/svn@17793 9fca08f4-af9d-4005-b8df-a31f2cc04f65