upgpkg: ddclient 3.8.0-3 Fix the stop function in the init script.
We cannot reliably use pidof to obtain the PID of ddclient, because
ddclient modifies the contents of argv[0] while it is running.
The default config file specifies a location of /var/run/ddclient.pid
for the PID file, so this should be a reliable way to get the PID
of a running ddclient.


git-svn-id: file:///srv/repos/svn-community/svn@51174 9fca08f4-af9d-4005-b8df-a31f2cc04f65