upgpkg: bsd-games 2.17-13
Fix FS#30554; add patches from Debian.

The "hunt" program contained an incorrect cast.
A comparison is made against the return value of ntohl.  The original
code casts the righthand operand to unsigned long, but
ntohl actually returns uint32_t.

Also, we need to add a NULL pointer check to another function in
hunt.c, to prevent a possible NULL pointer dereference, which leads
to a segfault.  This is completely unrelated to the issue in FS#30554.

The patches are courtesy of Debian.



git-svn-id: file:///srv/repos/svn-community/svn@73815 9fca08f4-af9d-4005-b8df-a31f2cc04f65