upgpkg: mysql-workbench 6.3.6-2
Back in 2013 we decided to replace upstream MySQL by Oracle with MariaDB [0].
Sadly forks tend to diverge over time and MariaDB is no longer a drop in
replacement. Take a look at the numerous patches for details.
Thus I decided to bundle upstream libraries:

* mysql (libmysqlclient.so, mysql, mysqldump)
* mysql-connector-c++ (libmysqlcppconn.so)
* gdal (libgdal.so) (Arch's gdal is linked against MariaDB's libmysqlclient
  and makes versions clash)

As a result we have more build instructions but fewer (quick'n'dirty) patches.

We can drop mysql-connector-c++ to AUR once this hits [community].

[0] https://www.archlinux.org/news/mariadb-replaces-mysql-in-repositories/

git-svn-id: file:///srv/repos/svn-community/svn@156178 9fca08f4-af9d-4005-b8df-a31f2cc04f65