upgpkg: keepassxc 2.3.0-1
New enabled feature: KeePassXC-Browser and SSH Agent. Now we enable all feature.
Note that KeePassHTTP is expected to be removed in next 2.4 version.
New dependencies: argon2 for KDBX4 format, curl for favicons retrieval.
libgcrypt was a transient dependency, so added it.

We don't build i686 anymore, so remove test exception.


git-svn-id: file:///srv/repos/svn-community/svn@300178 9fca08f4-af9d-4005-b8df-a31f2cc04f65