FS#58797 Fix makedepends build cycles
It is now possible to build haskell-colour with --nocheck in order to bootstrap
things. This results in different binary products depending on whether the
tests were built, but these seem to change quite often depending on many other
factors so one would be advised to rebuild all reverse dependencies for every
rebuild of haskell-colour either way.



git-svn-id: file:///srv/repos/svn-community/svn@343013 9fca08f4-af9d-4005-b8df-a31f2cc04f65