LLVM 7.0 rebuild + target master
The beignet project is now in maintenance mode only (it has been superseeded by
Neo/compute-runtime), and even new releases are unlikely.
However, this project might still be usefull for those with IvyBridge & Haswell
GPUs as those will not be supported by Neo/compute-runtime.

Most of the commits in the master branches are fixes toward LLVM 6 & 7
compatibility, everything else is bug fixes pushed by Debian, so following
master (but pinning the commit to know where we are) is a good idea now.


git-svn-id: file:///srv/repos/svn-community/svn@384126 9fca08f4-af9d-4005-b8df-a31f2cc04f65