upgpkg: python-cram 0.7-4
fix testsuites using the wrong version of python. The tests use a python runner
to re-execute a bunch of shellscripts that then re-re-execute the module
itself. And those tests try to find which version of python to use based on the
default python -- or $PYTHON. This then also shipped python3 bytecode in the
python2 package. AGHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHH.

git-svn-id: file:///srv/repos/svn-community/svn@432014 9fca08f4-af9d-4005-b8df-a31f2cc04f65