upgpkg: python-autobahn 19.6.2-1
The missing dependency issue [1] is fixed in this version. However, the fixed
setup.py specifies cryptography>=2.7, and python-cryptography 2.7-1 is still
in [testing].

Do not move this package to [community] until python-cryptography 2.7-1 is
moved to [extra]!

community/python-autobahn is stuck at 19.6.1. In that version setup.py does not
declare a dependency on cryptography, so there will not be
pkg_resources.VersionConflict. Also, according to results of the test suite,
cryptography 2.6 is most likely compatible.

[1] https://github.com/crossbario/autobahn-python/issues/1205

git-svn-id: file:///srv/repos/svn-community/svn@477447 9fca08f4-af9d-4005-b8df-a31f2cc04f65