upgpkg: python-nose-random 1.0.0-3
* Update homepage and source URLs; upstream project has been transferred to a
  different GitHub organization.
* Disable check(). See details below.

`examples/tests.py` is more like a demonstration rather than a unit test
suite. `failling_test()` fails, but with a small probability (5%), so I
didn't notice that.

There are more issues. When a test fails, another exception occurs if the
interpreter is Python 3:

    AttributeError: 'AssertionError' object has no attribute 'message'

This is because PEP 352 [1] removed the message attribute.

As python-nose is dead [2], and will be broken as of Python 3.10 as it
uses imp [3], I will not put efforts into fixing python-nose-random. In
the future, I will ask if jsondiff can switch to pytest or nose2.

[1] https://www.python.org/dev/peps/pep-0352/
[2] https://github.com/nose-devs/nose/issues/1092
[3] https://www.python.org/dev/peps/pep-0594/

git-svn-id: file:///srv/repos/svn-community/svn@510837 9fca08f4-af9d-4005-b8df-a31f2cc04f65