sigil: do not refer to as a WYSIWYG editor
It was never really... recommended... to use Book View. And now that Book View
is gone, sigil is not even conceivable to describe as a WYSIWYG editor anymore.

Update the pkgdesc to use the current official tagline, with a minor tweak to
refer to both EPUB revisions (sigil can handle both, so let's advertise both).



git-svn-id: file:///srv/repos/svn-community/svn@510950 9fca08f4-af9d-4005-b8df-a31f2cc04f65