upgpkg: iwd 1.0-3
We enable static builds with an ELL shipped inside of the tarball again,
because upstream *strongly* recommends to not use ELL as shared library.  ELL
is not on a stable release branch yet.

This will mitigate shared library errors for the future (at least as long ELL
is not considered stable).

git-svn-id: file:///srv/repos/svn-community/svn@520998 9fca08f4-af9d-4005-b8df-a31f2cc04f65