upgpkg: python-cleo 0.6.8-1
Abort! Abort! Abort! This is a downpkg, actually. The original version of cleo
I tried to package is actually too new for poetry, so we need an old one. I
escaped needing to add an epoch by a hair, because I didn't db-update this yet.
DON'T LOOK AT ME LIKE THAT IT'S FINE.


git-svn-id: file:///srv/repos/svn-community/svn@530403 9fca08f4-af9d-4005-b8df-a31f2cc04f65