upgpkg: calibre 4.4.0-1: upstream release + python3
The python3 port is pretty far along now, and I've tested out parallel installs
in calibre-git on the AUR for a while now. Let's get this better tested and
hopefully find more bugs while preparing for official python3 support. Since I
anticipate people will at a minimum want to be able to run plugins, this shares
common code in a -common package, and comes with a simple script to switch the
single set of conflicting files (the /usr/bin entry points) between the python2
and python3 versions. Users can install either or both.

git-svn-id: file:///srv/repos/svn-community/svn@533159 9fca08f4-af9d-4005-b8df-a31f2cc04f65