This was cloned from "ruby-unf" for incorrect reasons, see FS#65283
The existing package has good reasons to exceptionally package a revision later
than the latest stable release, and we do not just go around creating packages
literally called -beta, nor do we volunteer other people to be the maintainer
of said package.

Let's wait for consensus on how to resolve the bug report rather than trying to
pull a fait accompli as soon as someone else expresses a dissenting opinion.


git-svn-id: file:///srv/repos/svn-community/svn@562628 9fca08f4-af9d-4005-b8df-a31f2cc04f65