keybase: check if service is running in keybase-gui
... and don't just activate the service in this case. This ensures repeated
calls to keybase-gui will cause the app to foreground.

This is of greater importance now, since electron 8 dropped support for XEmbed
tray icons, which means some desktops might not see a Keybase tray icon, which
in turn means that foregrounding the application is hard to do with just a
service. Now, frantic clicking on the Keybase desktop file while trying to
figure out where the application went to, will actually have a visible effect
and cause it to come out from where it was "hiding".


git-svn-id: file:///srv/repos/svn-community/svn@594070 9fca08f4-af9d-4005-b8df-a31f2cc04f65