upgpkg: helm-synth 0.9.0-8: Adding helm-synth 0.9.0.
This package replaces helm (the synthesizer) to be able to move kubernetes' helm to the repositories.
Additionally helm-synth is fixed to be buildable against gcc > 9.1.0 and removes the unneeded
requirement to link against libcurl.so (which was introduced for no reason).
The vst plugin is now also built, which makes aur/helm-vst obsolete.

git-svn-id: file:///srv/repos/svn-community/svn@603504 9fca08f4-af9d-4005-b8df-a31f2cc04f65