upgpkg: weechat-matrix 0.1.0-2: install to site-packages
Improve the packaging thanks to dvzrv and eschwartz: the Python module is now
properly installed to site-pages, which has the advantage of tracked dependency
versions as well as pregenerated bytecode. The only thing left in
/usr/share/weechat/python/ is the WeeChat script that needs to be symlinked to
WEECHAT_HOME by the user. It was renamed to weechat-matrix.py because using the
name matrix.py leads to a circular import error.

As a result users will have to update their symbolic links after the upgrade,
this is documented by a post_upgrade message.


git-svn-id: file:///srv/repos/svn-community/svn@608346 9fca08f4-af9d-4005-b8df-a31f2cc04f65