python-webtest: better PYTHONPATH hack for adding currentdir
The current specification doesn't work at all, since there is no
architecture-specific build/lib.* directory. But the fallback to $PYTHONPATH
has the effect, if PYTHONPATH is empty, of adding the current directory to the
PYTHONPATH, and this causes the egg-info to be picked up by pkg_resources and
avoid a DistributionNotFound.



git-svn-id: file:///srv/repos/svn-community/svn@609002 9fca08f4-af9d-4005-b8df-a31f2cc04f65