upgpkg: lib32-jack2 1.9.14-3: Rebuilding without mixed mode.
Upstream's mixed mode is still broken and also more or less useless.
It was left in the PKGBUILD for testing purposes so the last version accidentally was built with it.
This is tracked in FS#66164.
Only pulling in direct library dependencies and build dependencies (db/lib32-db is not required).

git-svn-id: file:///srv/repos/svn-community/svn@612558 9fca08f4-af9d-4005-b8df-a31f2cc04f65