upgpkg: llpp 31.r41.g233b1a0-1: mupdf staticlibs rebuild
As usual, this entails updating to the latest git master so it will actually
build. On the plus side, we can drop a patch for strange -include usage as it's
rendered obsolete by upstream development, which doesn't strangely -include
anymore.

Also drop unneeded dependency on desktop stuff we aren't responsible for, and
switch glu dep to libgl, since we don't need glu, but we do need libgl which it
drew in.


git-svn-id: file:///srv/repos/svn-community/svn@625225 9fca08f4-af9d-4005-b8df-a31f2cc04f65