upgpkg: libolm 3.1.4-3: add libprovides/deps, avoid compiling in check()
To test the library auto-generated by CFFI without installing the Python
package, the "build/lib.<platform>-<pyver>" (e.g. "build/lib.linux-x86_64-3.8")
directory must be added to the Python path. This is more elegant than the
previous version of running the CFFI helper "olm_build.py" to recreate the
library in the current directory where it can be found directly by Python: it
avoids compiling in check() and makes sure that we test exactly the same library
that is going to get installed by the package.


git-svn-id: file:///srv/repos/svn-community/svn@625731 9fca08f4-af9d-4005-b8df-a31f2cc04f65