upgpkg: android-tools 30.0.0-1
As usual a new update requires a lot of changes to the build script. This is
an inevitable price that we pay for partial mirroring of the Android build
system to reuse host-level software components.

Upstream android platform tools use libstdc++ from LLVM project. Arch does not
have this package in the repo thus we use libc++ for now.

There are minor source-level incompatibilities that are most likely caused by
the fact that we use libc++. Libc++ seems more strict when it comes to
including headers for particular POSIX functions. We have to patch "./core"
project and add required "#includes".

Android uses c++2a standard extensively including features that have been
added to compilers just recently. The latest release of android-tools
requires GCC10.

Android-tools requires boringssl. We compile libssl.a and libcrypo.a as static
libraries and then use it for linking with adb/fastboot. System openssl
libraries are not needed in this case.

Android uses an mdns library that is API incompatible with mdns we have at Arch.
Disable mdns-specific code for now.


git-svn-id: file:///srv/repos/svn-community/svn@626247 9fca08f4-af9d-4005-b8df-a31f2cc04f65