upgpkg: ibm-sw-tpm2 1628-2: increase NV memory
Version 1628 enabled RSA 3072 objects, which need more storage space, without
increasing the available memory. As a result, only three objects can be
persisted (the previous version 1563 could store up to five objects), which
causes issues with the current master branch of the tpm2-pkcs11 test suite. The
memory size is a configurable option that must be set a compile time.


git-svn-id: file:///srv/repos/svn-community/svn@636676 9fca08f4-af9d-4005-b8df-a31f2cc04f65