upgpkg: tpm2-pkcs11 1.2.0-3: always enable unit tests for reproducibility
If unit tests are enabled, the macro UNIT_TESTING is defined, which changes the
code of src/lib/twist.c and thus the built library. If reproducers try to build
the package with --nocheck to save resources, the package contents will not
match. Unlike integration tests, enabling unit tests unconditionally only
requires a makedepends on cmocka, which is tiny.

Also update commit hashes of the patches to build with GCC 10 and ibm-sw-tpm2
1628 to the official merged commits that will be part of the next release.

git-svn-id: file:///srv/repos/svn-community/svn@657811 9fca08f4-af9d-4005-b8df-a31f2cc04f65