upgpkg: thunderbird-extension-enigmail 2.1.7-1: (final) upstream release
This is the last version of Enigmail, Thunderbird 78 will not support the
extension any more, offering native OpenPGP support instead. By default the
extension displays a warning to this effect, which might lead some users to
perform unsupported partial upgrades. Hence we disable this warning and replace
it by a versioned dependency on thunderbird<78 to prevent accidental
installations of Enigmail in combination with the next Thunderbird release.


git-svn-id: file:///srv/repos/svn-community/svn@657815 9fca08f4-af9d-4005-b8df-a31f2cc04f65