upgpkg: java-commons-lang 3.10-2: replace strip-nondeterminism by project.build.outputTimestamp
This is the officially supported way for reproducible builds with Apache Maven,
see https://maven.apache.org/guides/mini/guide-reproducible-builds.html
There are some additional sources of unreproducibilty in the form of manifest
headers added by maven-bundle-plugin. These can be removed by a suitable plugin
configuration, which has been suggested to upstream and backported to the
current release for this package.

git-svn-id: file:///srv/repos/svn-community/svn@658819 9fca08f4-af9d-4005-b8df-a31f2cc04f65