upgpkg: calibre 5.0.0-1: upstream release
Upstream is now fully migrated to python3 and python2 is no longer supported or
compatible. Good riddance! We now move back to a single package which replaces
all previous alternatives.

Moves from python-sip (4.x) to sip5 and sip-builder to build.

Also note one test is being ignored since it crashes in Qt, but only in
makechrootpkg/nspawn, and works fine on the host.


git-svn-id: file:///srv/repos/svn-community/svn@711955 9fca08f4-af9d-4005-b8df-a31f2cc04f65