upgpkg: jack2 1.9.15-1: Upgrade to 1.9.15.
Add all direct dependencies in depends.
Add all available sodeps in package() and the respective packages to makedepends.
Add zita-ajbridge to optdepends, as it is the only way to achieve a multi-card setup, which jack does natively.
Switch to signed tags for upstream verification, as the maintainer is now signing tags.

git-svn-id: file:///srv/repos/svn-community/svn@724095 9fca08f4-af9d-4005-b8df-a31f2cc04f65