upgpkg: python-wxpython 4.0.7.2-3: remove unneeded pathlib2 makedep, fix tests
pathlib2 is never used, on python3 it uses pathlib and on python2 it vendors
pathlib2 for the build. We only care about the former.

Tests are bonkers, the builtin test runner reports pytest output but
unconditionally return success, make this explicit by the PKGBUILD pattern
|| echo "warning goes here"

Also the tests never ran due to passing pytest --boxed, but this needed
pytest-xdist and is now a no-op, because the behavior is now in pytest-forked
as --forked. Nevertheless, it makes for clearer errors if we get segfaults, so
reimplement this properly.

git-svn-id: file:///srv/repos/svn-community/svn@749942 9fca08f4-af9d-4005-b8df-a31f2cc04f65