upgpkg: python-poetry 1.1.4-1: upstream release
lots of tweaks to get it to build sanely, including backporting a patch that
fixes the "trying to write to /foo during tests" thing.

The highlight of this release is that the "poetry-core" PyPI module is now the
build backend, as a lot of stuff got refactored out to there. This package is
therefore now recent enough that python-poetry-core no longer conflicts with
it.


git-svn-id: file:///srv/repos/svn-community/svn@842057 9fca08f4-af9d-4005-b8df-a31f2cc04f65