upgpkg: lib32-libsndfile 1.0.31-1: Upgrade to 1.0.31.
Switch to new github upstream for url and source.
Add new PGP public keys, that replace/ extend the existing:
https://github.com/libsndfile/libsndfile/issues/609
The original author transferred the project to a github
organization, but left the chain of trust broken.
Switch to cmake as build system.
Add all available sodeps in package() and the respective packages
in makedepends.

git-svn-id: file:///srv/repos/svn-community/svn@845531 9fca08f4-af9d-4005-b8df-a31f2cc04f65