upgpkg: nextcloud 21.0.0-6: Rebuild to make file installation more generic.
Make the find call to install files of the web app more generic.
Remove the user change by environment variable of the /usr/bin/occ command,
as it caused issues due to sudo not picking it up properly.
Change the group ownership of directories below /usr/share/webapps/nextcloud
back to root in the .install file (with nextcloud < 21.0.0 the files were
recursively assigned to the http group, but this is not necessary).

git-svn-id: file:///srv/repos/svn-community/svn@867612 9fca08f4-af9d-4005-b8df-a31f2cc04f65