upgpkg: nextcloud 21.0.0-6: Rebuild to make file installation more generic.
Make the find call to install files of the web app more generic.
Change the group ownership of directories below /usr/share/webapps/nextcloud
back to root in the .install file (with nextcloud < 21.0.0 the files were
recursively assigned to the http group, but this is not necessary).
Install writable app and data dir with 770 file mode.
Amend the .install message to not reference the previous (wrongly introduced)
location of the configuration file.
Amend the /usr/bin/occ command to allow overrides with NEXTCLOUD_USER,
NEXTCLOUD_PHP and NEXTCLOUD_PHP_CONFIG (for user, php and php config
respectively).

git-svn-id: file:///srv/repos/svn-community/svn@867619 9fca08f4-af9d-4005-b8df-a31f2cc04f65