upgpkg: uwsgi 2.0.19.1-9: Rebuild to add php >= 8.0.0 support.
Update pkgdesc to upstream description sentence.
Update url to current readthedocs page.
Reduce makedepends to essentials for the built plugins (and add php-embed).
Remove the optdepends on systemd (it is not needed).
Move backup array to uwsgi package() function (the config is part of that
package).
Remove ancient conflicts against python-uwsgi.
Do not use outdated uswsgiplugin.py for the systemd_logger plugin.
Remove the unused uwsgi_fix_rpath.patch.
Replace the ancient uwsgi_trick_chroot.patch with a plugin_build_dir entry in
archlinux.ini.
Add backported patch of upstream's fix for php >= 8.0.0
(https://github.com/unbit/uwsgi/pull/2288).
Rename tmpfilesd to uwsgi.tmpfiles for consistency.
Rename ruby >= 2.0 patch, so that it carries the version of uwsgi when it was
introduced.
Make applying of patches in prepare() more generic.
Replace the UWSGICONFIG_LUAPC override in build() with a sed call in prepare().
Simplify modification of ruby plugin configuration (renaming).
Create a copy of the php plugin and build it against php7-embed, while the
default is built
against the default php-embed.
Replace the separate plugin build calls with entries in archlinux.ini.
Use verbose output for the build.
Unify the style of the various package() functions and simplify the install
calls.
Add all direct dependencies and all available sodeps in the package()
functions.
Update maintainer info.

git-svn-id: file:///srv/repos/svn-community/svn@870867 9fca08f4-af9d-4005-b8df-a31f2cc04f65