upgpkg: buildbot 2.10.1-2; let systemd manage the buildbot user and its $HOME
Before this change, buildbot and buildbot-worker systemd services run
with the `buildbot` user and the working directory /var/lib/buildbot as
per upstream unit files, while the packages do not create the user and
the directory and users need to do that manually. This commit fills the
gap.

Note that this should not be a breaking change as the user should have
been running with the specified user and directory. If a user overrides
systemd configurations, they need to also override tmpfiles and sysusers
configurations from now on.

This commit also fixes a test failure with python-moto 2.0.0.


git-svn-id: file:///srv/repos/svn-community/svn@876399 9fca08f4-af9d-4005-b8df-a31f2cc04f65