upgpkg: python-pillow 8.2.0-2: make libraqm a hard dependency
It pulls in freetype and thus makes ImageFont a mandatory feature. But without
it, it is no longer possible to install freetype to enable ImageFont, (since
libraqm is no longer dlopened, but linked to at load time) and furthermore
there's no logical connection to the dependency, so it's obscure to try
depending on it.

But ultimately, the biggest problem is that the current version does not even
work sanely if the dependencies are not installed. It is supposed to raise an
ImportError telling you the _imagingft module is not installed, instead it
dumps an inscrutable stacktrace from the guts of the 'packaging' module.


git-svn-id: file:///srv/repos/svn-community/svn@924664 9fca08f4-af9d-4005-b8df-a31f2cc04f65