busybox: pass CONFIG_STATIC_LIBGCC
Dunno why it got added in the off state during the latest update. It's probably
harmless given we build statically via musl anyway, but it defaults to on for a
good reason so let's keep it like that in the conf.

This used to be passed regardless (hardcoded in Makefile.flags), but the latest
version added an option for the people who really really don't like it, so they
don't have to use it. Omitting it here wasn't a conscious or principled choice...

git-svn-id: file:///srv/repos/svn-community/svn@934296 9fca08f4-af9d-4005-b8df-a31f2cc04f65