python-tomlkit: remove dephell workaround for building
Awkward bootstrap model, frozen development. It's not necessarily needed
anymore, now that current versions of poetry use setuptools.

The other reason to use it was due to handling of the tests/ directory, but the
very latest version of tomlkit finally uses a new poetry build-backend feature
to specify that the tests directory should be included in sdist, but not added
as an installable module.

git-svn-id: file:///srv/repos/svn-community/svn@939702 9fca08f4-af9d-4005-b8df-a31f2cc04f65