upgpkg: keycloak 14.0.0-3: pin java runtime for exeuction
Dependency constraint won't solve the issue as the default java can still be
any arbitrary java. In cases where we need to enforce a specific java we need
to hand select a java target and patch the runtime execution to actually force
use that specific version.

As keycloak shell scripts are out of tree and somewhere deep down the rabbit
hole lets use a patch file that we dyinamicall adjust and execute during
packaging.


git-svn-id: file:///srv/repos/svn-community/svn@965731 9fca08f4-af9d-4005-b8df-a31f2cc04f65