upgpkg: osquery 5.0.1-1
osquery devendorizing activity did not go anywhere [1].

Maitaining a large patch and porting it to the new osquery releases is a tremendous work.
Rather spending time on the huge patch we switch to more upstream-like static build.

The cmake build pulls all the dependencies and builds it itself.
The main difference is that we do not use libc++ (as libLLVM does not work well with it)
and instead use the default library from gcc.

[1] https://github.com/osquery/osquery/issues/6286

git-svn-id: file:///srv/repos/svn-community/svn@1020720 9fca08f4-af9d-4005-b8df-a31f2cc04f65