upgpkg: py3status 3.39-2: add optdepends on python-pytz
The clock module has a hard dependency on the deprecated python-pytz because it
uses pytz.country_timezones() and therefore cannot be easily replaced with the
newer built-in zoneinfo module. This is unlikely to change in version 3.x of
py3status: https://github.com/ultrabug/py3status/pull/1992#issuecomment-758504036

The clock module is also the only module to use the external tzlocal module, so
edit the description of the optional dependency accordingly.


git-svn-id: file:///srv/repos/svn-community/svn@1032950 9fca08f4-af9d-4005-b8df-a31f2cc04f65