upgpkg: cni-plugins 1.0.1-2: Rebuild to add provides array.
As netavark is now in the repositories, we have two providers of container-network-stack,
so we add container-network-stack to provides.
Disable LTO, as the plugins seem not compatible with it:
https://github.com/containernetworking/plugins/issues/706
Update the info around why we install to /opt/cni/bin as well.
Remove unneeded explicit installation of /opt/cni/bin.
Update maintainer info.

git-svn-id: file:///srv/repos/svn-community/svn@1135253 9fca08f4-af9d-4005-b8df-a31f2cc04f65