Remove polkit as dependency
We can leverage seatd to also obtain elevated system privileges instead of
polkit. As such, we cannot require polkit since there is an alternative.

Note that using seatd instead of polkit means that a user must be a part of the
"seat" group: Otherwise, the user will not be able to achieve the appropriate
privileged actions.

git-svn-id: file:///srv/repos/svn-community/svn@1209451 9fca08f4-af9d-4005-b8df-a31f2cc04f65