Pin the hash of the tag object instead of the tag name for git sources
Tags can be force pushed upstream, so relying on the tag name is not enough to
guarantee integrity of the sources. Instead we can pin the SHA-1 hash of the
tag object (obtained using "git rev-parse"), which would change in the event of
a force push.

Also add a pkgver() function to avoid accidentally bumping only $pkgver instead
of updating $_tag.

git-svn-id: file:///srv/repos/svn-community/svn@1214769 9fca08f4-af9d-4005-b8df-a31f2cc04f65