upgpkg: tang 11-2: install key directory (once again)
The upgrade to tang 11-1 accidentally removed the creation of the key directory
(because "Z" in the tmpfiles.d configuration file only alters the permission of
existing directories, it does not create them if they do not exist). This means
that tang does not work out of the box because it cannot automatically create
the required keys.

Fix this by restoring the accidentally removed "mkdir" call. Creating the empty
directory as part of the package instead of only through the tmpfiles.d
configuration file means that it can be properly tracked by
"pacman -Qo /var/db/tang/".

git-svn-id: file:///srv/repos/svn-community/svn@1222852 9fca08f4-af9d-4005-b8df-a31f2cc04f65