upgpkg: python-cleo 1.0.0a5-1: Upgrade to 1.0.0a5.
Bump to prerelease version because upstream poetry released 1.2.0 with a cleo
>= 1.0.0a5 dependency (https://github.com/python-poetry/poetry/issues/6353) and
cleo likely will not see a stable release before poetry 1.3.0 is out
(https://github.com/python-poetry/cleo/issues/246).
As the likelyhood of running into instabilities due to other poetry
dependencies when downgrading to poetry < 1.2.0 is higher than when upgrading
only cleo to an unstable version, upgrading cleo to an unstable version is the
lesser of the two evils.

Update upstream url.
Remove unnecessary quotes.
Switch to PEP517.

git-svn-id: file:///srv/repos/svn-community/svn@1290920 9fca08f4-af9d-4005-b8df-a31f2cc04f65