upgpkg: audacity 1:3.2.1-1: Upgrade to 3.2.1.
Switch to ffmpeg for additional import/export capabilities.
Use system provided vst3sdk to build vst3 plugin host support and -DNDEBUG to
CFLAGS and CXXFLAGS to be able to build it.
Add new makedepends on mpg123 and wavpack and simplify the list of
dependencies.
Switch to the auto-generated tarball, as we are using none of the submodules
anyways and the custom tarball is now weird:
https://github.com/audacity/audacity/issues/3811
Remove unneeded patches and fixes.
Build in release mode using the custom cmake variable AUDACITY_BUILD_LEVEL, as
the build is otherwise marked as alpha.

git-svn-id: file:///srv/repos/svn-community/svn@1329016 9fca08f4-af9d-4005-b8df-a31f2cc04f65