upgpkg: consul 1.13.3-1 | bump to 1.13.3, switch to VCS source
Switching to the git source will be required for building the UI as there are some hardcoded git commands in the build suite for it, that expect to be run from within a git checkout.
We're not currently building the UI (yet), but there is no harm in switching to the VCS source with a pinned commit.
(The release tarballs on github weren't signed either.)


git-svn-id: file:///srv/repos/svn-community/svn@1335440 9fca08f4-af9d-4005-b8df-a31f2cc04f65